memo.js内で行っていたデーターベース関係処理をデータベース処理ファイルに移動
