+++ 04.class/memo.js
@@ -1,40 +1,13 @@
 #! /usr/bin/env node
 
 import minimist from "minimist";
-import { existsSync } from "node:fs";
-import { writeFile } from "node:fs/promises";
 import { DatabaseConnector } from "./lib/database_connector.js";
 import { NoteList } from "./lib/note_list.js";
 
 class MemoApp {
-  constructor() {
-    this.filePath = "./db/memo.db";
-    this.dbConnector = null;
-  }
-
-  static async initialize() {
-    const memoApp = new MemoApp();
-
-    if (!existsSync(memoApp.filePath)) {
-      await writeFile(memoApp.filePath, "");
-    }
-
-    memoApp.dbConnector = new DatabaseConnector(memoApp.filePath);
-    try {
-      await memoApp.dbConnector.createTable();
-    } catch (err) {
-      if (err instanceof Error && err.code === "SQLITE_ERROR") {
-        console.error(err);
-      } else {
-        throw err;
-      }
-    }
-
-    return memoApp;
-  }
-
   async run() {
-    const noteList = new NoteList(this.dbConnector);
+    const noteDb = await DatabaseConnector.initialize();
+    const noteList = new NoteList(noteDb);
     const options = minimist(process.argv.slice(2));
     if (options.l) {
       await noteList.seeAllTitles();
@@ -48,9 +21,5 @@ class MemoApp {
   }
 }
 
-const main = async () => {
-  const memoApp = await MemoApp.initialize();
-  memoApp.run();
-};
-
-main();
+const main = new MemoApp();
+main.run();
